+++ AuthPermissions.SupportCode/ShardingServices/GetDatabaseForNewTenant.cs
@@ -18,11 +18,13 @@ public class GetDatabaseForNewTenant : IGetDatabaseForNewTenant
 
     /// <summary>
     /// This will look for a database for a new tenant.
-    /// If the hasOwnDb is true, then it will find an empty database
+    /// If the hasOwnDb is true, then it will find an empty database,
+    /// otherwise it will look for database containing multiple tenants
     /// </summary>
-    /// <param name="hasOwnDb"></param>
+    /// <param name="hasOwnDb">If true the tenant needs its own database. False means it shares a database.</param>
+    /// <param name="region">If not null this provides geographic information to pick the nearest database server.</param>
     /// <returns>Status with the DatabaseInfoName, or error if it can't find a database to work with</returns>
-    public async Task<IStatusGeneric<string>> FindBestDatabaseInfoNameAsync(bool hasOwnDb)
+    public async Task<IStatusGeneric<string>> FindBestDatabaseInfoNameAsync(bool hasOwnDb, string region)
     {
         var status = new StatusGenericHandler<string>();