+++ Example3.MvcWebApp.IndividualAccounts/Controllers/HomeController.cs
@@ -3,8 +3,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System.Diagnostics;
 using System.Threading.Tasks;
+using AuthPermissions.SupportCode.AddUsersServices;
 using Example3.InvoiceCode.Dtos;
 using Example3.InvoiceCode.Services;
+using Example3.MvcWebApp.IndividualAccounts.PermissionsCode;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 
@@ -40,17 +42,14 @@ namespace Example3.MvcWebApp.IndividualAccounts.Controllers
         [AllowAnonymous]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> CreateTenant(CreateTenantDto data,
-            [FromServices] IUserRegisterInviteService userRegisterInvite,
-            [FromServices] SignInManager<IdentityUser> signInManager)
+        public async Task<IActionResult> CreateTenant([FromServices] ISignInAndCreateTenant userRegisterInvite,
+            AddNewTenantDto data, string password, bool isPersistent)
         {
-            var status = await userRegisterInvite.AddUserAndNewTenantAsync(data);
+            var status = await userRegisterInvite.AddUserAndNewTenantAsync(data, 
+                CreateTenantSettings.TenantSetupData);
             if (status.HasErrors)
                 return RedirectToAction(nameof(ErrorDisplay),
                     new { errorMessage = status.GetAllErrors() });
-
-            //User has been successfully registered so now we need to log them in
-            await signInManager.SignInAsync(status.Result, isPersistent: false);
 
             return RedirectToAction(nameof(Index),
                 new { message = status.Message });
@@ -65,17 +64,13 @@ namespace Example3.MvcWebApp.IndividualAccounts.Controllers
         [AllowAnonymous]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> AcceptInvite(AcceptInviteDto data,
-            [FromServices] IUserRegisterInviteService userRegisterInvite,
-            [FromServices] SignInManager<IdentityUser> signInManager)
+        public async Task<ActionResult> AcceptInvite([FromServices] IInviteNewUserService inviteUserServiceService,
+            string verify, string email, string password, bool isPersistent)
         {
-            var status = await userRegisterInvite.AcceptUserJoiningATenantAsync(data.Email, data.Password, data.Verify);
+            var status = await inviteUserServiceService.AddUserViaInvite(verify, email, password, isPersistent);
             if (status.HasErrors)
                 return RedirectToAction(nameof(ErrorDisplay),
                     new { errorMessage = status.GetAllErrors() });
-
-            //User has been successfully registered so now we need to log them in
-            await signInManager.SignInAsync(status.Result, isPersistent: false);
 
             return RedirectToAction(nameof(Index),
                 new { message = status.Message });