+++ Example3.MvcWebApp.IndividualAccounts/Models/AcceptInviteDto.cs
@@ -5,7 +5,8 @@ namespace Example3.MvcWebApp.IndividualAccounts.Models;
 
 public class AcceptInviteDto
 {
+    public string Verify { get; set; }
     public string Email { get; set; }
     public string Password { get; set; }
-    public string Verify { get; set; }
+    public bool IsPersistent { get; set; }
 }