+++ Example3.MvcWebApp.IndividualAccounts/Models/InviteUserDto.cs
@@ -1,19 +1,17 @@
-﻿// Copyright (c) 2021 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
+﻿// Copyright (c) 2022 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
 namespace Example3.MvcWebApp.IndividualAccounts.Models;
 
 public class InviteUserDto
 {
-    public InviteUserDto(string email, string tenantName, string url)
+    public InviteUserDto(string message, string url)
     {
-        Email = email;
-        TenantName = tenantName;
+        Message = message;
         Url = url;
     }
 
-    public string Email { get; }
-    public string TenantName { get;  }
+    public string Message { get; }
     public string Url { get;  }
 
 }